refactor: drop cycle & types in protocol
