+++ test/unit/formatPaymentInfo.test.ts
@@ -1,7 +1,7 @@
 import { assert } from 'chai'
 
 import { formatPaymentInfo } from '../../src/services/basePayId'
-import { AddressDetailsType } from '../../src/types/publicAPI'
+import { AddressDetailsType } from '../../src/types/protocol'
 
 describe('Base PayID - formatPaymentInfo()', function (): void {
   it('Returns CryptoAddressDetails & FiatAddressDetails for addressDetailsTypes when formatting array with multiple AddressInformation', function () {
